requirements.txt
aiofiles==23.1.0
aiomysql==0.1.1
aiosqlite==0.17.0
asyncmy==0.2.5
atlastk==0.13.2
certifi==2022.12.7
charset-normalizer==3.0.1
coloredlogs==15.0.1
docopt==0.6.2
executor==23.2
fasteners==0.18
greenlet==2.0.2
httptools==0.5.0
humanfriendly==10.0
idna==3.4
iso8601==1.1.0
mayim==1.0.1
multidict==6.0.4
pip==23.0
pipreqs==0.4.11
property-manager==3.0
pydantic==1.10.4
PyMySQL==1.0.2
pypika-tortoise==0.1.6
pyreadline3==3.4.1
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
sanic-ext==22.12.0
sanic-routing==22.8.0
sanic==22.12.0
setuptools==67.2.0
six==1.16.0
SQLAlchemy==2.0.3
tortoise-orm==0.19.3
typing-extensions==4.4.0
urllib3==1.26.14
verboselogs==1.7
websockets==10.4
wheel==0.38.4
yarg==0.1.9
tortoise==0.1.1
ujson==5.7.0